+++ CHANGELOG.md
@@ -1,14 +1,22 @@
 ## Version 0.4.3-SNAPSHOT
 
-This is a work-in-progress containing minor changes from v0.4.2.
+This is a *minor release* that aims to be fully compatible with previous v0.4.x releases, while fixing bugs.
 
-v0.4.3 will be released if there are enough bugs to fix... otherwise the changes will be incorporated into v0.5.0.
+### Enhancements
+* Support Bio-Formats memoization (again) (https://github.com/qupath/qupath/issues/1236)
+  * To use this, set the 'Bio-Formats memoization time' in QuPath's preferences
 
 ### Bugs fixed
 * Opening the same image in multiple viewers results in detections being wrongly shown in both (https://github.com/qupath/qupath/issues/1217)
 * 'Transform annotations' can't be applied to selected objects when double-clicking to end the transform (https://github.com/qupath/qupath/issues/1231)
 * Wrong script generated for 'Add intensity features' with some tile shapes (https://github.com/qupath/qupath/issues/1227)
-* 'Tile classifications to annotations' broken in v0.4.2 (https://github.com/qupath/qupath/issues/1226)
+* 'Tile classifications to annotations' is broken (https://github.com/qupath/qupath/issues/1226)
+
+### Dependency updates
+* Bio-Formats 6.12.0
+  * See https://bio-formats.readthedocs.io/en/v6.12.0/about/whats-new.html
+* Groovy 4.0.9
+* JFreeSVG 5.0.5
 
 
 ## Version 0.4.2